// src/generated/models/Cr9cd_gamesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9cd_gamescr9cd_kind = {
  100000000: 'Game',
  100000001: 'Event'
} as const;
export type Cr9cd_gamescr9cd_kind = keyof typeof Cr9cd_gamescr9cd_kind;
export const Cr9cd_gamescr9cd_status = {
  100000000: 'Scheduled',
  100000001: 'Transfer Pending',
  100000002: 'Completed',
  100000003: 'Cancelled'
} as const;
export type Cr9cd_gamescr9cd_status = keyof typeof Cr9cd_gamescr9cd_status;
export const Cr9cd_gamesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9cd_gamesstatecode = keyof typeof Cr9cd_gamesstatecode;
export const Cr9cd_gamesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9cd_gamesstatuscode = keyof typeof Cr9cd_gamesstatuscode;

export interface Cr9cd_gamesBase {
  cr9cd_description?: string;
  cr9cd_game_date?: string;
  cr9cd_gameid: string;
  cr9cd_kind?: Cr9cd_gamescr9cd_kind;
  cr9cd_name?: string;
  cr9cd_notes?: string;
  cr9cd_opponent?: string;
  cr9cd_premium_score?: number;
  cr9cd_promotions?: string;
  "cr9cd_Season@odata.bind": string;
  cr9cd_status?: Cr9cd_gamescr9cd_status;
  cr9cd_title?: string;
  cr9cd_total_seats?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9cd_gamesstatecode;
  statuscode?: Cr9cd_gamesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9cd_games extends Cr9cd_gamesBase {
  cr9cd_kindname?: string;
  cr9cd_seasonname?: string;
  cr9cd_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr9cd_season?: object;
  _cr9cd_season_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}