// src/generated/models/Cr9cd_notificationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9cd_notificationscr9cd_type = {
  100000000: 'Reminder',
  100000001: 'Announcement',
  100000002: 'Availability',
  100000003: 'Game Link'
} as const;
export type Cr9cd_notificationscr9cd_type = keyof typeof Cr9cd_notificationscr9cd_type;
export const Cr9cd_notificationscr9cd_channel = {
  100000000: 'Email',
  100000001: 'SMS',
  100000002: 'In-App'
} as const;
export type Cr9cd_notificationscr9cd_channel = keyof typeof Cr9cd_notificationscr9cd_channel;
export const Cr9cd_notificationscr9cd_audience = {
  100000000: 'Everyone',
  100000001: 'Sales Team',
  100000002: 'Employees',
  100000003: 'Season-Ticket Holders'
} as const;
export type Cr9cd_notificationscr9cd_audience = keyof typeof Cr9cd_notificationscr9cd_audience;
export const Cr9cd_notificationscr9cd_status = {
  100000000: 'Draft',
  100000001: 'Scheduled',
  100000002: 'Sent'
} as const;
export type Cr9cd_notificationscr9cd_status = keyof typeof Cr9cd_notificationscr9cd_status;
export const Cr9cd_notificationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9cd_notificationsstatecode = keyof typeof Cr9cd_notificationsstatecode;
export const Cr9cd_notificationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9cd_notificationsstatuscode = keyof typeof Cr9cd_notificationsstatuscode;

export interface Cr9cd_notificationsBase {
  cr9cd_audience?: Cr9cd_notificationscr9cd_audience;
  cr9cd_channel?: Cr9cd_notificationscr9cd_channel;
  "cr9cd_Game@odata.bind"?: string;
  cr9cd_message?: string;
  cr9cd_name: string;
  cr9cd_notificationid: string;
  cr9cd_recipient_count?: number;
  cr9cd_scheduled_at?: string;
  cr9cd_sent_at?: string;
  cr9cd_status?: Cr9cd_notificationscr9cd_status;
  cr9cd_type?: Cr9cd_notificationscr9cd_type;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9cd_notificationsstatecode;
  statuscode?: Cr9cd_notificationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9cd_notifications extends Cr9cd_notificationsBase {
  cr9cd_audiencename?: string;
  cr9cd_channelname?: string;
  cr9cd_gamename?: string;
  cr9cd_statusname?: string;
  cr9cd_typename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr9cd_game?: object;
  _cr9cd_game_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}